Enable PluginData for users
