+++ lib/auth/auth_with_roles.go
@@ -1229,6 +1229,13 @@ func (a *ServerWithRoles) GetPluginData(ctx context.Context, filter types.Plugin
 			}
 		}
 		return a.authServer.GetPluginData(ctx, filter)
+
+	case types.KindUser:
+		// TODO: Put correct auth checks here
+		if err := a.action(apidefaults.Namespace, types.KindUserPluginData, types.VerbRead); err != nil {
+			return nil, trace.Wrap(err)
+		}
+		return a.authServer.GetPluginData(ctx, filter)
 	default:
 		return nil, trace.BadParameter("unsupported resource kind %q", filter.Kind)
 	}
@@ -1246,6 +1253,13 @@ func (a *ServerWithRoles) UpdatePluginData(ctx context.Context, params types.Plu
 			}
 		}
 		return a.authServer.UpdatePluginData(ctx, params)
+	case types.KindUser:
+		// TODO: Put correct auth checks here
+		if err := a.action(apidefaults.Namespace, types.KindUserPluginData, types.VerbUpdate); err != nil {
+			return trace.Wrap(err)
+		}
+		return a.authServer.UpdatePluginData(ctx, params)
+
 	default:
 		return trace.BadParameter("unsupported resource kind %q", params.Kind)
 	}
